+++ api/routes/base/types.ts
@@ -1,0 +1,7 @@
+import type { AutoRouterType } from "itty-router";
+import type { Services } from "../../services/install";
+
+export type RoutesRegisterHandler = (
+  router: AutoRouterType,
+  installServices: ({ env }: { env: Env }) => Services,
+) => void;
