+++ api/tests/server.test.ts
@@ -5,7 +5,6 @@ import { InteractionType } from "discord-api-types/v10";
 import type * as HaloInfiniteApi from "halo-infinite-api";
 import { HaloInfiniteClient, StaticXstsTicketTokenSpartanTokenProvider } from "halo-infinite-api";
 import { installFakeServicesWith } from "../services/fakes/services";
-import type { DatabaseService } from "../services/database/database";
 import { Server } from "../server";
 import { getCommands } from "../commands/commands";
 import { aFakeEnvWith } from "../base/fakes/env.fake";
@@ -56,334 +55,6 @@ describe("Server", () => {
       expect(res.status).toBe(404);
       const text = await res.text();
       expect(text).toBe("Not Found.");
-    });
-  });
-
-  describe("POST /auth/logout", () => {
-    it("returns 200, revokes the server session, and clears the session cookie", async () => {
-      let deleteUserSessionSpy!: MockInstance<DatabaseService["deleteUserSession"]>;
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "validateSession").mockResolvedValue({
-          sessionId: "session-123",
-          userId: "user-123",
-          accessToken: "access-token",
-          refreshToken: "refresh-token",
-          expiresAt: Date.now() + 3600000,
-          isExpired: false,
-        });
-        deleteUserSessionSpy = vi.spyOn(services.databaseService, "deleteUserSession").mockResolvedValue();
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/logout", {
-        method: "POST",
-        headers: {
-          Cookie: "auth-session=valid-token",
-        },
-      });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(200);
-      const body = await res.json<{ success: boolean }>();
-      expect(body).toEqual({ success: true });
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-      const setCookie = res.headers.get("Set-Cookie");
-      expect(setCookie).toContain("auth-session=");
-      expect(setCookie).toContain("Max-Age=0");
-      expect(deleteUserSessionSpy).toHaveBeenCalledWith("session-123");
-    });
-
-    it("returns 200 and clears the session cookie when server-side invalidation fails", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "invalidateSession").mockRejectedValue(new Error("D1 unavailable"));
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/logout", { method: "POST" });
-      const res = (await server.router.fetch(req, env)) as Response;
-
-      expect(res.status).toBe(200);
-      const body = await res.json<{ success: boolean }>();
-      expect(body).toEqual({ success: true });
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-      const setCookie = res.headers.get("Set-Cookie");
-      expect(setCookie).toContain("auth-session=");
-      expect(setCookie).toContain("Max-Age=0");
-    });
-
-    it("returns 500 with error message when clearSessionCookie throws", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "clearSessionCookie").mockImplementation(() => {
-          throw new Error("Cookie error");
-        });
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/logout", { method: "POST" });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(500);
-      const body = await res.json<{ error: string }>();
-      expect(body).toEqual({ error: "Logout failed" });
-    });
-  });
-
-  describe("GET /auth/session", () => {
-    it("returns 401 with authenticated false when no session cookie is present", async () => {
-      const req = new Request("http://localhost/auth/session", { method: "GET" });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(401);
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-      const body = await res.json<{ authenticated: boolean }>();
-      expect(body).toEqual({ authenticated: false });
-    });
-
-    it("adds credentialed CORS headers for allowed origins", async () => {
-      const req = new Request("http://localhost/auth/session", {
-        method: "GET",
-        headers: {
-          Origin: env.PAGES_URL,
-        },
-      });
-      const res = (await server.router.fetch(req, env)) as Response;
-
-      expect(res.headers.get("Access-Control-Allow-Origin")).toBe(env.PAGES_URL);
-      expect(res.headers.get("Access-Control-Allow-Credentials")).toBe("true");
-      expect(res.headers.get("Vary")).toBe("Origin");
-    });
-
-    it("returns 401 with expired flag when session is expired", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "validateSession").mockResolvedValue({
-          sessionId: "session-123",
-          userId: "user-123",
-          accessToken: "access-token",
-          refreshToken: undefined,
-          expiresAt: Date.now() - 1000,
-          isExpired: true,
-        });
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/session", { method: "GET" });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(401);
-      const body = await res.json<{ authenticated: boolean; expired: boolean }>();
-      expect(body).toEqual({ authenticated: false, expired: true });
-      expect(res.headers.get("Set-Cookie")).toContain("auth-session=");
-      expect(res.headers.get("Set-Cookie")).toContain("Max-Age=0");
-    });
-
-    it("returns 200 with refreshed session info when access token is expired but refresh succeeds", async () => {
-      const refreshedExpiresAt = Date.now() + 3600000;
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "validateSession").mockResolvedValue({
-          sessionId: "session-123",
-          userId: "user-123",
-          accessToken: "expired-access-token",
-          refreshToken: "refresh-token",
-          expiresAt: Date.now() - 1000,
-          isExpired: true,
-        });
-        vi.spyOn(services.authService, "refreshSession").mockResolvedValue({
-          sessionId: "session-123",
-          userId: "user-123",
-          accessToken: "fresh-access-token",
-          refreshToken: "fresh-refresh-token",
-          expiresAt: refreshedExpiresAt,
-          issuedAt: Date.now(),
-        });
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-
-      const req = new Request("http://localhost/auth/session", { method: "GET" });
-      const res = (await server.router.fetch(req, env)) as Response;
-
-      expect(res.status).toBe(200);
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-      const body = await res.json<{ authenticated: boolean; userId: string; expiresAt: number }>();
-      expect(body).toEqual({ authenticated: true, userId: "user-123", expiresAt: refreshedExpiresAt });
-    });
-
-    it("returns 200 with user info when session is valid", async () => {
-      const expiresAt = Date.now() + 3600000;
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "validateSession").mockResolvedValue({
-          sessionId: "session-123",
-          userId: "user-123",
-          accessToken: "access-token",
-          refreshToken: undefined,
-          expiresAt,
-          isExpired: false,
-        });
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/session", { method: "GET" });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(200);
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-      const body = await res.json<{ authenticated: boolean; userId: string; expiresAt: number }>();
-      expect(body).toEqual({ authenticated: true, userId: "user-123", expiresAt });
-    });
-
-    it("returns 500 with error message when validateSession throws", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "validateSession").mockRejectedValue(new Error("Session error"));
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/session", { method: "GET" });
-      const res = (await server.router.fetch(req, env)) as Response;
-      expect(res.status).toBe(500);
-      const body = await res.json<{ error: string }>();
-      expect(body).toEqual({ error: "Failed to retrieve session" });
-    });
-  });
-
-  describe("GET /auth/microsoft/start", () => {
-    it("adds credentialed CORS headers for allowed origins", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "generateAuthorizationUrl").mockResolvedValue({
-          url: new URL("https://login.microsoftonline.com/test"),
-          state: "state-123",
-          codeVerifier: "verifier-123",
-        });
-        vi.spyOn(services.authService, "setPkceStateCookie").mockResolvedValue();
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/microsoft/start", {
-        method: "GET",
-        headers: {
-          Origin: env.PAGES_URL,
-        },
-      });
-      const res = (await server.router.fetch(req, env)) as Response;
-
-      expect(res.status).toBe(200);
-      expect(res.headers.get("Access-Control-Allow-Origin")).toBe(env.PAGES_URL);
-      expect(res.headers.get("Access-Control-Allow-Credentials")).toBe("true");
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
-    });
-  });
-
-  describe("GET /auth/microsoft/callback", () => {
-    it("sets a long-lived session cookie separate from access token expiry", async () => {
-      vi.useFakeTimers();
-      try {
-        vi.setSystemTime(new Date("2026-05-17T23:05:18.409Z"));
-
-        const accessTokenExpiresAt = Date.now() + 3600 * 1000;
-        const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-          const services = installFakeServicesWith({ env });
-          vi.spyOn(services.authService, "handleCallback").mockResolvedValue({
-            sessionId: "session-123",
-            userId: "user-123",
-            accessToken: "access-token",
-            refreshToken: "refresh-token",
-            expiresAt: accessTokenExpiresAt,
-            issuedAt: Date.now(),
-          });
-          vi.spyOn(services.authService, "createSessionToken").mockResolvedValue("signed-session-token");
-          return services;
-        });
-        server = new Server({
-          router: AutoRouter(),
-          installServices: localInstallServices,
-          getCommands,
-        });
-        const req = new Request("http://localhost/auth/microsoft/callback?code=code-123&state=state-123", {
-          method: "GET",
-          headers: {
-            Origin: env.PAGES_URL,
-          },
-        });
-
-        const res = (await server.router.fetch(req, env)) as Response;
-
-        expect(res.status).toBe(200);
-        const body = await res.json<{ success: boolean; userId: string }>();
-        expect(body).toEqual({ success: true, userId: "user-123" });
-        const setCookie = res.headers.get("Set-Cookie");
-        expect(setCookie).toContain("auth-session=signed-session-token");
-        expect(setCookie).toContain("Max-Age=2592000");
-
-        const expiresAtMatch = setCookie?.match(/auth-session=[^]*?Expires=([^;]+GMT)/);
-        expect(expiresAtMatch).not.toBeNull();
-        const cookieExpiresAt = Date.parse(expiresAtMatch?.[1] ?? "");
-        expect(cookieExpiresAt).toBeGreaterThan(accessTokenExpiresAt);
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it("returns a generic authentication error when callback handling fails", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.authService, "handleCallback").mockRejectedValue(
-          new Error("upstream microsoft response body"),
-        );
-        return services;
-      });
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-      const req = new Request("http://localhost/auth/microsoft/callback?code=code-123&state=state-123", {
-        method: "GET",
-        headers: {
-          Origin: env.PAGES_URL,
-        },
-      });
-      const res = (await server.router.fetch(req, env)) as Response;
-
-      expect(res.status).toBe(400);
-      const body = await res.json<{ error: string }>();
-      expect(body).toEqual({ error: "Authentication failed" });
-      expect(res.headers.get("Access-Control-Allow-Origin")).toBe(env.PAGES_URL);
-      expect(res.headers.get("Access-Control-Allow-Credentials")).toBe("true");
-      expect(res.headers.get("Cache-Control")).toBe("no-store");
     });
   });
